Use rethought message content view and nested types
