+++ Package.swift
@@ -220,38 +220,38 @@ var streamChatSourcesExcluded: [String] { [
 ] }
 
 var streamChatUIFilesExcluded: [String] { [
-    "ChatMessageList/RETHINK/MessageContentView_Tests.swift",
-    "ChatMessageList/RETHINK/__Snapshots__/MessageContentView_Tests/test_appearance.default-light-flipped-bubble-continuousBubble-text-quotedMessage-reactions.png",
-    "ChatMessageList/RETHINK/__Snapshots__/MessageContentView_Tests/test_appearance.default-light-flipped-bubble-continuousBubble-text-quotedMessage-threadInfo-reactions.png",
-    "ChatMessageList/RETHINK/__Snapshots__/MessageContentView_Tests/test_appearanceCustomization_usingSubclassing.small-dark.png",
-    "ChatMessageList/RETHINK/__Snapshots__/MessageContentView_Tests/test_appearance.default-light-flipped-bubble-continuousBubble-metadata-text-quotedMessage-threadInfo-error-reactions.png",
-    "ChatMessageList/RETHINK/__Snapshots__/MessageContentView_Tests/test_appearanceCustomization_usingUIConfig.small-dark.png",
-    "ChatMessageList/RETHINK/__Snapshots__/MessageContentView_Tests/test_appearance.default-light-flipped-bubble-continuousBubble-metadata-text-quotedMessage-threadInfo-reactions.png",
-    "ChatMessageList/RETHINK/__Snapshots__/MessageContentView_Tests/test_appearance.default-light-flipped-bubble-continuousBubble-text-reactions.png",
-    "ChatMessageList/RETHINK/__Snapshots__/MessageContentView_Tests/test_appearance.default-light-flipped-bubble-continuousBubble-text-quotedMessage.png",
-    "ChatMessageList/RETHINK/__Snapshots__/MessageContentView_Tests/test_appearance.default-light-flipped-bubble-continuousBubble-text-quotedMessage-threadInfo-errorIndicator-reactions.png",
-    "ChatMessageList/RETHINK/__Snapshots__/MessageContentView_Tests/test_appearance.default-light-bubble-avatar-metadata-text-quotedMessage-reactions.png",
-    "ChatMessageList/RETHINK/__Snapshots__/MessageContentView_Tests/test_appearance.default-light-flipped-bubble-metadata-text-quotedMessage-reactions.png",
-    "ChatMessageList/RETHINK/__Snapshots__/MessageContentView_Tests/test_appearance.default-light-bubble-continuousBubble-avatarSizePadding-text.png",
-    "ChatMessageList/RETHINK/__Snapshots__/MessageContentView_Tests/test_appearance.default-light-bubble-continuousBubble-avatar-metadata-text-quotedMessage-threadInfo-reactions.png",
-    "ChatMessageList/RETHINK/__Snapshots__/MessageContentView_Tests/test_appearance.default-light-flipped-bubble-continuousBubble-text.png",
-    "ChatMessageList/RETHINK/__Snapshots__/MessageContentView_Tests/test_appearance.default-light-flipped-bubble-metadata-text-reactions.png",
-    "ChatMessageList/RETHINK/__Snapshots__/MessageContentView_Tests/test_appearance.default-light-bubble-avatar-metadata-text.png",
-    "ChatMessageList/RETHINK/__Snapshots__/MessageContentView_Tests/test_appearance.default-light-bubble-continuousBubble-avatarSizePadding-text-reactions.png",
-    "ChatMessageList/RETHINK/__Snapshots__/MessageContentView_Tests/test_appearance.default-light-flipped-bubble-continuousBubble-text-quotedMessage-threadInfo-error-reactions.png",
-    "ChatMessageList/RETHINK/__Snapshots__/MessageContentView_Tests/test_appearance.default-light-flipped-bubble-continuousBubble-metadata-text-quotedMessage-threadInfo-errorIndicator-reactions.png",
-    "ChatMessageList/RETHINK/__Snapshots__/MessageContentView_Tests/test_appearance.default-light-bubble-continuousBubble-avatarSizePadding-text-quotedMessage.png",
-    "ChatMessageList/RETHINK/__Snapshots__/MessageContentView_Tests/test_appearance.default-light-flipped-bubble-metadata-text-quotedMessage.png",
-    "ChatMessageList/RETHINK/__Snapshots__/MessageContentView_Tests/test_appearance.default-light-bubble-avatar-metadata-text-quotedMessage.png",
-    "ChatMessageList/RETHINK/__Snapshots__/MessageContentView_Tests/test_appearanceCustomization_usingUIConfig.extraExtraExtraLarge-light.png",
-    "ChatMessageList/RETHINK/__Snapshots__/MessageContentView_Tests/test_appearance.default-light-bubble-continuousBubble-avatarSizePadding-text-quotedMessage-reactions.png",
-    "ChatMessageList/RETHINK/__Snapshots__/MessageContentView_Tests/test_appearanceCustomization_usingSubclassing.extraExtraExtraLarge-light.png",
-    "ChatMessageList/RETHINK/__Snapshots__/MessageContentView_Tests/test_appearanceCustomization_usingUIConfig.default-light.png",
-    "ChatMessageList/RETHINK/__Snapshots__/MessageContentView_Tests/test_appearance.default-light-flipped-bubble-metadata-text.png",
-    "ChatMessageList/RETHINK/__Snapshots__/MessageContentView_Tests/test_appearance.default-light-bubble-continuousBubble-avatarSizePadding-text-quotedMessage-threadInfo-reactions.png",
-    "ChatMessageList/RETHINK/__Snapshots__/MessageContentView_Tests/test_appearance.default-light-bubble-avatar-metadata-text-reactions.png",
-    "ChatMessageList/RETHINK/__Snapshots__/MessageContentView_Tests/test_appearanceCustomization_usingSubclassing.default-light.png",
     "ChatMessageList/ChatMessageListCollectionViewLayout_Tests.swift",
+    "ChatMessageList/ChatMessage/ChatMessageContentView_Tests.swift",
+    "ChatMessageList/ChatMessage/__Snapshots__/ChatMessageContentView_Tests/test_appearance.default-light-flipped-bubble-continuousBubble-text-quotedMessage-reactions.png",
+    "ChatMessageList/ChatMessage/__Snapshots__/ChatMessageContentView_Tests/test_appearance.default-light-flipped-bubble-continuousBubble-text-quotedMessage-threadInfo-reactions.png",
+    "ChatMessageList/ChatMessage/__Snapshots__/ChatMessageContentView_Tests/test_appearanceCustomization_usingSubclassing.small-dark.png",
+    "ChatMessageList/ChatMessage/__Snapshots__/ChatMessageContentView_Tests/test_appearance.default-light-flipped-bubble-continuousBubble-metadata-text-quotedMessage-threadInfo-error-reactions.png",
+    "ChatMessageList/ChatMessage/__Snapshots__/ChatMessageContentView_Tests/test_appearanceCustomization_usingUIConfig.small-dark.png",
+    "ChatMessageList/ChatMessage/__Snapshots__/ChatMessageContentView_Tests/test_appearance.default-light-flipped-bubble-continuousBubble-metadata-text-quotedMessage-threadInfo-reactions.png",
+    "ChatMessageList/ChatMessage/__Snapshots__/ChatMessageContentView_Tests/test_appearance.default-light-flipped-bubble-continuousBubble-text-reactions.png",
+    "ChatMessageList/ChatMessage/__Snapshots__/ChatMessageContentView_Tests/test_appearance.default-light-flipped-bubble-continuousBubble-text-quotedMessage.png",
+    "ChatMessageList/ChatMessage/__Snapshots__/ChatMessageContentView_Tests/test_appearance.default-light-flipped-bubble-continuousBubble-text-quotedMessage-threadInfo-errorIndicator-reactions.png",
+    "ChatMessageList/ChatMessage/__Snapshots__/ChatMessageContentView_Tests/test_appearance.default-light-bubble-avatar-metadata-text-quotedMessage-reactions.png",
+    "ChatMessageList/ChatMessage/__Snapshots__/ChatMessageContentView_Tests/test_appearance.default-light-flipped-bubble-metadata-text-quotedMessage-reactions.png",
+    "ChatMessageList/ChatMessage/__Snapshots__/ChatMessageContentView_Tests/test_appearance.default-light-bubble-continuousBubble-avatarSizePadding-text.png",
+    "ChatMessageList/ChatMessage/__Snapshots__/ChatMessageContentView_Tests/test_appearance.default-light-bubble-continuousBubble-avatar-metadata-text-quotedMessage-threadInfo-reactions.png",
+    "ChatMessageList/ChatMessage/__Snapshots__/ChatMessageContentView_Tests/test_appearance.default-light-flipped-bubble-continuousBubble-text.png",
+    "ChatMessageList/ChatMessage/__Snapshots__/ChatMessageContentView_Tests/test_appearance.default-light-flipped-bubble-metadata-text-reactions.png",
+    "ChatMessageList/ChatMessage/__Snapshots__/ChatMessageContentView_Tests/test_appearance.default-light-bubble-avatar-metadata-text.png",
+    "ChatMessageList/ChatMessage/__Snapshots__/ChatMessageContentView_Tests/test_appearance.default-light-bubble-continuousBubble-avatarSizePadding-text-reactions.png",
+    "ChatMessageList/ChatMessage/__Snapshots__/ChatMessageContentView_Tests/test_appearance.default-light-flipped-bubble-continuousBubble-text-quotedMessage-threadInfo-error-reactions.png",
+    "ChatMessageList/ChatMessage/__Snapshots__/ChatMessageContentView_Tests/test_appearance.default-light-flipped-bubble-continuousBubble-metadata-text-quotedMessage-threadInfo-errorIndicator-reactions.png",
+    "ChatMessageList/ChatMessage/__Snapshots__/ChatMessageContentView_Tests/test_appearance.default-light-bubble-continuousBubble-avatarSizePadding-text-quotedMessage.png",
+    "ChatMessageList/ChatMessage/__Snapshots__/ChatMessageContentView_Tests/test_appearance.default-light-flipped-bubble-metadata-text-quotedMessage.png",
+    "ChatMessageList/ChatMessage/__Snapshots__/ChatMessageContentView_Tests/test_appearance.default-light-bubble-avatar-metadata-text-quotedMessage.png",
+    "ChatMessageList/ChatMessage/__Snapshots__/ChatMessageContentView_Tests/test_appearanceCustomization_usingUIConfig.extraExtraExtraLarge-light.png",
+    "ChatMessageList/ChatMessage/__Snapshots__/ChatMessageContentView_Tests/test_appearance.default-light-bubble-continuousBubble-avatarSizePadding-text-quotedMessage-reactions.png",
+    "ChatMessageList/ChatMessage/__Snapshots__/ChatMessageContentView_Tests/test_appearanceCustomization_usingSubclassing.extraExtraExtraLarge-light.png",
+    "ChatMessageList/ChatMessage/__Snapshots__/ChatMessageContentView_Tests/test_appearanceCustomization_usingUIConfig.default-light.png",
+    "ChatMessageList/ChatMessage/__Snapshots__/ChatMessageContentView_Tests/test_appearance.default-light-flipped-bubble-metadata-text.png",
+    "ChatMessageList/ChatMessage/__Snapshots__/ChatMessageContentView_Tests/test_appearance.default-light-bubble-continuousBubble-avatarSizePadding-text-quotedMessage-threadInfo-reactions.png",
+    "ChatMessageList/ChatMessage/__Snapshots__/ChatMessageContentView_Tests/test_appearance.default-light-bubble-avatar-metadata-text-reactions.png",
+    "ChatMessageList/ChatMessage/__Snapshots__/ChatMessageContentView_Tests/test_appearanceCustomization_usingSubclassing.default-light.png",
     "ChatMessageList/MessageComposer/ChatMessageComposerSuggestionsViewController_Tests.swift",
     "ChatMessageList/MessageComposer/ChatMessageComposerCommandCollectionViewCell_Tests.swift",
     "ChatMessageList/MessageComposer/ChatMessageComposerImageAttachmentsView_Tests.swift",
@@ -459,17 +459,11 @@ var streamChatUIFilesExcluded: [String] { [
     "MessageActionsPopup/ChatMessagePopupVC_Tests.swift",
     "MessageActionsPopup/ChatMessageActionsVC_Tests.swift",
     "MessageActionsPopup/__Snapshots__/ChatMessagePopupVC_Tests/test_appearanceCustomization_usingSubclassing.small-dark.png",
-    "MessageActionsPopup/__Snapshots__/ChatMessagePopupVC_Tests/test_defaultAppearance.default-light.png",
     "MessageActionsPopup/__Snapshots__/ChatMessagePopupVC_Tests/test_appearanceCustomization_usingAppearance.extraExtraExtraLarge-light.png",
-    "MessageActionsPopup/__Snapshots__/ChatMessagePopupVC_Tests/test_defaultAppearance_when_largeMessageViewFrame.small-dark.png",
-    "MessageActionsPopup/__Snapshots__/ChatMessagePopupVC_Tests/test_defaultAppearance_when_largeMessageViewFrame.default-light.png",
-    "MessageActionsPopup/__Snapshots__/ChatMessagePopupVC_Tests/test_defaultAppearance_when_largeMessageViewFrame.extraExtraExtraLarge-light.png",
     "MessageActionsPopup/__Snapshots__/ChatMessagePopupVC_Tests/test_defaultAppearance_when_largeLongMessage.default-light.png",
     "MessageActionsPopup/__Snapshots__/ChatMessagePopupVC_Tests/test_appearanceCustomization_usingSubclassing.extraExtraExtraLarge-light.png",
     "MessageActionsPopup/__Snapshots__/ChatMessagePopupVC_Tests/test_defaultAppearance_when_largeLongMessage.extraExtraExtraLarge-light.png",
     "MessageActionsPopup/__Snapshots__/ChatMessagePopupVC_Tests/test_appearanceCustomization_usingAppearance.default-light.png",
-    "MessageActionsPopup/__Snapshots__/ChatMessagePopupVC_Tests/test_defaultAppearance.small-dark.png",
-    "MessageActionsPopup/__Snapshots__/ChatMessagePopupVC_Tests/test_defaultAppearance.extraExtraExtraLarge-light.png",
     "MessageActionsPopup/__Snapshots__/ChatMessagePopupVC_Tests/test_appearanceCustomization_usingAppearance.small-dark.png",
     "MessageActionsPopup/__Snapshots__/ChatMessagePopupVC_Tests/test_appearanceCustomization_usingSubclassing.default-light.png",
     "MessageActionsPopup/__Snapshots__/ChatMessagePopupVC_Tests/test_defaultAppearance_when_largeLongMessage.small-dark.png",
